backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
fastapi
uvicorn
python-dotenv
httpx
